+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the eve-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Fixed
+- Added forgotten pagination support for /v3/characters/{character_id}/assets/.
+
 ## [4.1.1] - 2019-08-17
 ### Fixed
 - Errors in typings and forgotten string types.